Fix ESLint failures in webpack test and source files
- Auto-fix prettier formatting in JobInvocationActions.js,
  JobInvocationToolbarButtons.js, and three test files
- Split TemplateInvocation.test.js: move polling describe block to new
  TemplateInvocationPolling.test.js to stay under max-lines limit and
  eliminate no-shadow violations (store variable)
- Name anonymous function assigned to Date.prototype.toLocaleString in
  MainInformation.test.js to satisfy func-names rule
