+++ webpack/JobInvocationDetail/__tests__/TemplateInvocation.test.js
@@ -1,9 +1,7 @@
 import React from 'react';
 import configureMockStore from 'redux-mock-store';
-import { createStore, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
 import { Provider } from 'react-redux';
-import { render, screen, fireEvent, act } from '@testing-library/react';
+import { render, screen, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 import * as api from 'foremanReact/redux/API';
 import * as selectors from '../JobInvocationSelectors';
@@ -247,199 +245,5 @@ describe('TemplateInvocation', () => {
       );
       expect(postCall.key).toBe('ABORT_TASK');
     });
-  });
-});
-
-describe('TemplateInvocation polling', () => {
-  const noop = () => {};
-  const reducer = (state = {}) => state;
-  const makeStore = () => createStore(reducer, applyMiddleware(thunk));
-
-  const pollingProps = {
-    hostID: '1',
-    jobID: '1',
-    isInTableView: false,
-    isExpanded: true,
-    hostName: 'example-host',
-    hostProxy: { name: 'example-proxy', href: '#' },
-    showOutputType: { stderr: true, stdout: true, debug: true },
-    setShowOutputType: noop,
-    showTemplatePreview: false,
-    setShowTemplatePreview: noop,
-    showCommand: false,
-    setShowCommand: noop,
-  };
-
-  let apiGetSpy;
-
-  beforeEach(() => {
-    jest.useFakeTimers({ legacyFakeTimers: true });
-    selectors.selectTemplateInvocationStatus.mockImplementation(() => () =>
-      'RESOLVED'
-    );
-    selectors.selectTemplateInvocation.mockImplementation(() => () =>
-      mockTemplateInvocationResponse
-    );
-    apiGetSpy = jest
-      .spyOn(api.APIActions, 'get')
-      .mockImplementation(({ handleSuccess }) => {
-        handleSuccess &&
-          handleSuccess({ data: { finished: false, auto_refresh: true } });
-        return { type: 'MOCK_GET' };
-      });
-  });
-
-  afterEach(() => {
-    jest.clearAllTimers();
-    jest.useRealTimers();
-    apiGetSpy.mockRestore();
-  });
-
-  it('fetches on mount when isExpanded is true', () => {
-    const store = makeStore();
-    render(
-      <Provider store={store}>
-        <TemplateInvocation {...pollingProps} isExpanded />
-      </Provider>
-    );
-    expect(apiGetSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it('does not fetch on mount when isExpanded is false', () => {
-    const store = makeStore();
-    render(
-      <Provider store={store}>
-        <TemplateInvocation {...pollingProps} isExpanded={false} />
-      </Provider>
-    );
-    expect(apiGetSpy).not.toHaveBeenCalled();
-  });
-
-  it('schedules next poll via setTimeout when auto_refresh is true and not finished', () => {
-    const store = makeStore();
-    render(
-      <Provider store={store}>
-        <TemplateInvocation {...pollingProps} />
-      </Provider>
-    );
-    expect(apiGetSpy).toHaveBeenCalledTimes(1);
-
-    act(() => jest.advanceTimersByTime(5000));
-    expect(apiGetSpy).toHaveBeenCalledTimes(2);
-  });
-
-  it('does not schedule next poll when finished is true', () => {
-    apiGetSpy.mockImplementation(({ handleSuccess }) => {
-      handleSuccess && handleSuccess({ data: { finished: true, auto_refresh: true } });
-      return { type: 'MOCK_GET' };
-    });
-    const store = makeStore();
-    render(
-      <Provider store={store}>
-        <TemplateInvocation {...pollingProps} />
-      </Provider>
-    );
-    expect(apiGetSpy).toHaveBeenCalledTimes(1);
-
-    act(() => jest.advanceTimersByTime(5000));
-    expect(apiGetSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it('does not schedule next poll when auto_refresh is false', () => {
-    apiGetSpy.mockImplementation(({ handleSuccess }) => {
-      handleSuccess && handleSuccess({ data: { finished: false, auto_refresh: false } });
-      return { type: 'MOCK_GET' };
-    });
-    const store = makeStore();
-    render(
-      <Provider store={store}>
-        <TemplateInvocation {...pollingProps} />
-      </Provider>
-    );
-    expect(apiGetSpy).toHaveBeenCalledTimes(1);
-
-    act(() => jest.advanceTimersByTime(5000));
-    expect(apiGetSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it('clears the polling timeout and sets cancelled on unmount', () => {
-    const store = makeStore();
-    const { unmount } = render(
-      <Provider store={store}>
-        <TemplateInvocation {...pollingProps} />
-      </Provider>
-    );
-    expect(apiGetSpy).toHaveBeenCalledTimes(1);
-
-    unmount();
-    act(() => jest.advanceTimersByTime(5000));
-    expect(apiGetSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it('cancels in-flight callback when unmounted before handleSuccess runs', () => {
-    let capturedHandleSuccess;
-    apiGetSpy.mockImplementation(({ handleSuccess }) => {
-      capturedHandleSuccess = handleSuccess;
-      return { type: 'MOCK_GET' };
-    });
-
-    const store = makeStore();
-    const { unmount } = render(
-      <Provider store={store}>
-        <TemplateInvocation {...pollingProps} />
-      </Provider>
-    );
-    expect(apiGetSpy).toHaveBeenCalledTimes(1);
-
-    unmount();
-
-    act(() => {
-      capturedHandleSuccess({ data: { finished: false, auto_refresh: true } });
-      jest.advanceTimersByTime(5000);
-    });
-
-    expect(apiGetSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it('re-fetches when isExpanded changes from false to true', () => {
-    const store = makeStore();
-    const { rerender } = render(
-      <Provider store={store}>
-        <TemplateInvocation {...pollingProps} isExpanded={false} />
-      </Provider>
-    );
-    expect(apiGetSpy).not.toHaveBeenCalled();
-
-    rerender(
-      <Provider store={store}>
-        <TemplateInvocation {...pollingProps} isExpanded />
-      </Provider>
-    );
-    expect(apiGetSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it('cancels existing poll and starts fresh when isExpanded changes', () => {
-    const store = makeStore();
-    const { rerender } = render(
-      <Provider store={store}>
-        <TemplateInvocation {...pollingProps} isExpanded />
-      </Provider>
-    );
-    expect(apiGetSpy).toHaveBeenCalledTimes(1);
-
-    rerender(
-      <Provider store={store}>
-        <TemplateInvocation {...pollingProps} isExpanded={false} />
-      </Provider>
-    );
-    act(() => jest.advanceTimersByTime(5000));
-    expect(apiGetSpy).toHaveBeenCalledTimes(1);
-
-    rerender(
-      <Provider store={store}>
-        <TemplateInvocation {...pollingProps} isExpanded />
-      </Provider>
-    );
-    expect(apiGetSpy).toHaveBeenCalledTimes(2);
   });
 });
